requirements.txt
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.0
asgiref==3.5.2
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
crcmod==1.7
cryptography==38.0.1
cycler==0.11.0
Django==4.1.1
djangorestframework==3.13.1
fonttools==4.37.1
idna==3.3
jmespath==0.10.0
joblib==1.1.0
kiwisolver==1.4.4
matplotlib==3.5.3
numpy==1.23.3
oss2==2.16.0
packaging==21.3
pandas==1.4.4
Pillow==9.2.0
pycparser==2.21
pycryptodome==3.15.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.1.0
urllib3==1.26.12
